+++ src/hdmf/backends/hdf5/h5tools.py
@@ -7,16 +7,16 @@ import logging
 import warnings
 
 from ...container import Container
-from ...utils import docval, getargs, popargs, call_docval_func, get_data_shape, get_docval
+from ...utils import docval, getargs, popargs, call_docval_func, get_data_shape
 from ...data_utils import AbstractDataChunkIterator
-from ...build import Builder, GroupBuilder, DatasetBuilder, LinkBuilder, BuildManager,\
-                     RegionBuilder, ReferenceBuilder, TypeMap, ObjectMapper
+from ...build import (Builder, GroupBuilder, DatasetBuilder, LinkBuilder, BuildManager, RegionBuilder,
+                      ReferenceBuilder, TypeMap, ObjectMapper)
 from ...spec import RefSpec, DtypeSpec, NamespaceCatalog, GroupSpec, NamespaceBuilder
 
-from .h5_utils import BuilderH5ReferenceDataset, BuilderH5RegionDataset, BuilderH5TableDataset,\
-                      H5DataIO, H5SpecReader, H5SpecWriter
+from .h5_utils import (BuilderH5ReferenceDataset, BuilderH5RegionDataset, BuilderH5TableDataset, H5DataIO,
+                       H5SpecReader, H5SpecWriter)
 
-from ..io import HDMFIO, UnsupportedOperation, WriteConfig
+from ..io import HDMFIO, UnsupportedOperation
 from ..warnings import BrokenLinkWarning
 
 ROOT_NAME = 'root'
@@ -265,6 +265,9 @@ class HDF5IO(HDMFIO):
               not opened already when calling this function.
 
         """
+
+        # TODO add deprecation warning. this function will no longer be supported. use export instead.
+
         source_filename, dest_filename, expand_external, expand_refs, expand_soft = getargs('source_filename',
                                                                                             'dest_filename',
                                                                                             'expand_external',
@@ -290,8 +293,11 @@ class HDF5IO(HDMFIO):
 
     @docval({'name': 'container', 'type': Container, 'doc': 'the Container object to write'},
             {'name': 'cache_spec', 'type': bool, 'doc': 'cache specification to file', 'default': True},
-            {'name': 'write_config', 'type': 'H5WriteConfig', 'doc': 'configuration settings for writing builders',
-             'default': None})
+            {'name': 'exhaust_dci', 'type': bool,
+             'doc': 'exhaust DataChunkIterators one at a time. If False, exhaust them concurrently',
+             'default': True},
+            {'name': 'link_data', 'type': bool,
+             'doc': 'If not specified otherwise link (True) or copy (False) HDF5 Datasets', 'default': True})
     def write(self, **kwargs):
         if self.__mode == 'r':
             raise UnsupportedOperation(("Cannot write to file %s in mode '%s'. "
@@ -329,10 +335,8 @@ class HDF5IO(HDMFIO):
          'doc': ('the Container object to export. If None, then the entire contents of the HDMFIO object will be '
                  'exported'),
          'default': None},
-        {'name': 'read_args', 'type': dict, 'doc': 'dict of arguments to use when calling read_io.read_builder',
-         'default': dict()},
-        {'name': 'write_config', 'type': 'H5WriteConfig', 'doc': 'configuration settings for writing builders',
-         'default': None},
+        {'name': 'write_args', 'type': dict,
+         'doc': 'arguments to use when calling write_builder with this HDf5IO object', 'default': dict()},
         {'name': 'cache_spec', 'type': bool, 'doc': 'whether to cache the specification to file',
          'default': True}
     )
@@ -348,17 +352,15 @@ class HDF5IO(HDMFIO):
                                        % (self.__path, self.__mode))
 
         src_io = getargs('src_io', kwargs)
-        write_config, cache_spec = popargs('write_config', 'cache_spec', kwargs)
-        if write_config is None:
-            write_config = H5WriteConfig()
+        write_args, cache_spec = popargs('write_args', 'cache_spec', kwargs)
 
-        if not isinstance(src_io, HDF5IO) and write_config.link_data:
+        if not isinstance(src_io, HDF5IO) and write_args.get('link_data', False):
             raise UnsupportedOperation("Cannot export from non-HDF5 backend %s to HDF5 with write_config "
                                        "link_data=True.")
 
-        write_config['export_source'] = src_io.source  # pass export_source=src_io.source to write_builder
+        write_args['export_source'] = src_io.source  # pass export_source=src_io.source to write_builder
         ckwargs = kwargs.copy()
-        ckwargs['write_config'] = write_config
+        ckwargs['write_args'] = write_args
         call_docval_func(super().export, ckwargs)
         if cache_spec:
             self.__cache_spec()
@@ -371,14 +373,21 @@ class HDF5IO(HDMFIO):
 
         Convenience function for export where the user does not need to instantiate a new HDF5IO object for writing.
 
+        Example usage:
+
+            old_io = HDF5IO('old.nwb', 'r')
+            new_io.export_io(io_args={'path': 'new_copy.nwb'}, src_io=old_io)
+
         See export for more details.
         """
         io_args = popargs('io_args', kwargs)
-        if 'mode' in io_args and io_args['mode'] != 'w':
-            raise ValueError("The 'mode' key in io_args must be 'w' if present.")
-        io_args['mode'] = 'w'
+        mode = io_args.get('mode', default='w')
+        if mode != 'w' and mode != 'w-':
+            raise UnsupportedOperation("The 'mode' key in io_args must be 'w' or 'w-' if present.")
+        io_args['mode'] = mode
 
-        if 'manager' in io_args and io_args['manager'] is not None:
+        manager = io_args.get('manager')
+        if manager:
             warnings.warn("The 'manager' key in io_args will be ignored.")
             io_args.remove('manager')
 
@@ -664,18 +673,22 @@ class HDF5IO(HDMFIO):
         self.__open_links = []
 
     @docval({'name': 'builder', 'type': GroupBuilder, 'doc': 'the GroupBuilder object representing the HDF5 file'},
-            {'name': 'write_config', 'type': 'H5WriteConfig', 'doc': 'configuration settings for writing builders',
-             'default': None})
+            {'name': 'link_data', 'type': bool,
+             'doc': 'If not specified otherwise link (True) or copy (False) HDF5 Datasets', 'default': True},
+            {'name': 'exhaust_dci', 'type': bool,
+             'doc': 'exhaust DataChunkIterators one at a time. If False, exhaust them concurrently',
+             'default': True},
+            {'name': 'export_source', 'type': str,
+             'doc': 'The source of the builders when exporting', 'default': None})
     def write_builder(self, **kwargs):
-        f_builder, write_config = getargs('builder', 'write_config', kwargs)
-        if write_config is None:
-            write_config = H5WriteConfig()
-        self.logger.debug("Writing GroupBuilder '%s' to path '%s' with write config: %s"
-                          % (f_builder.name, self.__path, write_config))
+        f_builder = popargs('builder', kwargs)
+        link_data, exhaust_dci, export_source = getargs('link_data', 'exhaust_dci', 'export_source', kwargs)
+        self.logger.debug("Writing GroupBuilder '%s' to path '%s' with kwargs=%s"
+                          % (f_builder.name, self.__path, kwargs))
         for name, gbldr in f_builder.groups.items():
-            self.write_group(self.__file, gbldr, write_config=write_config)
+            self.write_group(self.__file, gbldr, **kwargs)
         for name, dbldr in f_builder.datasets.items():
-            self.write_dataset(self.__file, dbldr, write_config=write_config)
+            self.write_dataset(self.__file, dbldr, **kwargs)
         for name, lbldr in f_builder.links.items():
             self.write_link(self.__file, lbldr)
         self.set_attributes(self.__file, f_builder.attributes)
@@ -830,13 +843,16 @@ class HDF5IO(HDMFIO):
 
     @docval({'name': 'parent', 'type': Group, 'doc': 'the parent HDF5 object'},
             {'name': 'builder', 'type': GroupBuilder, 'doc': 'the GroupBuilder to write'},
-            {'name': 'write_config', 'type': 'H5WriteConfig', 'doc': 'configuration settings for writing builders',
-             'default': None},
+            {'name': 'link_data', 'type': bool,
+             'doc': 'If not specified otherwise link (True) or copy (False) HDF5 Datasets', 'default': True},
+            {'name': 'exhaust_dci', 'type': bool,
+             'doc': 'exhaust DataChunkIterators one at a time. If False, exhaust them concurrently',
+             'default': True},
+            {'name': 'export_source', 'type': str,
+             'doc': 'The source of the builders when exporting', 'default': None},
             returns='the Group that was created', rtype='Group')
     def write_group(self, **kwargs):
-        parent, builder, write_config = getargs('parent', 'builder', 'write_config', kwargs)
-        if write_config is None:
-            write_config = H5WriteConfig()
+        parent, builder = popargs('parent', 'builder', kwargs)
         self.logger.debug("Writing GroupBuilder '%s' to parent group '%s'" % (builder.name, parent.name))
         if self.get_written(builder):
             group = parent[builder.name]
@@ -847,12 +863,12 @@ class HDF5IO(HDMFIO):
         if subgroups:
             for subgroup_name, sub_builder in subgroups.items():
                 # do not create an empty group without attributes or links
-                self.write_group(group, sub_builder, write_config=write_config)
+                self.write_group(group, sub_builder, **kwargs)
         # write all datasets
         datasets = builder.datasets
         if datasets:
             for dset_name, sub_builder in datasets.items():
-                self.write_dataset(group, sub_builder, write_config=write_config)
+                self.write_dataset(group, sub_builder, **kwargs)
         # write all links
         links = builder.links
         if links:
@@ -875,11 +891,9 @@ class HDF5IO(HDMFIO):
 
     @docval({'name': 'parent', 'type': Group, 'doc': 'the parent HDF5 object'},
             {'name': 'builder', 'type': LinkBuilder, 'doc': 'the LinkBuilder to write'},
-            {'name': 'write_config', 'type': 'H5WriteConfig', 'doc': 'configuration settings for writing builders',
-             'default': None},
             returns='the Link that was created', rtype='Link')
     def write_link(self, **kwargs):
-        parent, builder, write_config = getargs('parent', 'builder', 'write_config', kwargs)
+        parent, builder = getargs('parent', 'builder', kwargs)
         self.logger.debug("Writing LinkBuilder '%s' to parent group '%s'" % (builder.name, parent.name))
         if self.get_written(builder):
             return None
@@ -909,8 +923,13 @@ class HDF5IO(HDMFIO):
 
     @docval({'name': 'parent', 'type': Group, 'doc': 'the parent HDF5 object'},
             {'name': 'builder', 'type': DatasetBuilder, 'doc': 'the DatasetBuilder to write'},
-            {'name': 'write_config', 'type': 'H5WriteConfig', 'doc': 'configuration settings for writing builders',
-             'default': None},
+            {'name': 'link_data', 'type': bool,
+             'doc': 'If not specified otherwise link (True) or copy (False) HDF5 Datasets', 'default': True},
+            {'name': 'exhaust_dci', 'type': bool,
+             'doc': 'exhaust DataChunkIterators one at a time. If False, exhaust them concurrently',
+             'default': True},
+            {'name': 'export_source', 'type': str,
+             'doc': 'The source of the builders when exporting', 'default': None},
             returns='the Dataset that was created', rtype=Dataset)
     def write_dataset(self, **kwargs):  # noqa: C901
         """ Write a dataset to HDF5
@@ -918,16 +937,14 @@ class HDF5IO(HDMFIO):
         The function uses other dataset-dependent write functions, e.g,
         __scalar_fill__, __list_fill__ and __setup_chunked_dset__ to write the data.
         """
-        parent, builder, write_config = getargs('parent', 'builder', 'write_config', kwargs)
-        if write_config is None:
-            write_config = H5WriteConfig()
+        parent, builder = popargs('parent', 'builder', kwargs)
+        link_data, exhaust_dci, export_source = getargs('link_data', 'exhaust_dci', 'export_source', kwargs)
         self.logger.debug("Writing DatasetBuilder '%s' to parent group '%s'" % (builder.name, parent.name))
         if self.get_written(builder):
             self.logger.debug("    DatasetBuilder '%s' is already written" % builder.name)
             return None
         name = builder.name
         data = builder.data
-        link_data = write_config.link_data
         options = dict()   # dict with additional
         if isinstance(data, H5DataIO):
             options['io_settings'] = data.io_settings
@@ -943,7 +960,6 @@ class HDF5IO(HDMFIO):
         # The user provided an existing h5py dataset as input and asked to create a link to the dataset
         if isinstance(data, Dataset):
             data_filename = os.path.abspath(data.file.filename)
-            export_source = write_config.export_source
             if export_source is not None:
                 export_source = os.path.abspath(export_source)
             if link_data and (data_filename != export_source or parent.name != data.parent.name):
@@ -1091,7 +1107,7 @@ class HDF5IO(HDMFIO):
         elif len(attributes) > 0:
             pass
         self.__set_written(builder)
-        if write_config.exhaust_dci:
+        if exhaust_dci:
             self.__exhaust_dcis()
 
     @classmethod
@@ -1327,14 +1343,3 @@ class HDF5IO(HDMFIO):
         Return the HDF5 file mode. One of ("w", "r", "r+", "a", "w-", "x").
         """
         return self.__mode
-
-
-class H5WriteConfig(WriteConfig):
-
-    @docval(*get_docval(WriteConfig.__init__),
-            {'name': 'link_data', 'type': bool,
-             'doc': 'If not specified otherwise link (True) or copy (False) HDF5 Datasets', 'default': True})
-    def __init__(self, **kwargs):
-        link_data = popargs('link_data', kwargs)
-        call_docval_func(super().__init__, kwargs)
-        self['link_data'] = link_data